Fixed that mysql_upgrade doesn't give errors about mariadb.sys
The reason for this fix was that when I tried to run mysql_upgrade
at home to update an old 10.5 installation, mysql_upgrade failed
with warnings about mariadb.sys user not existing.

If the server was started with --skip-grants, there would be no warnings
from mysql_upgrade, but in some cases running mysql_upgrade again could
produce new warnings.

The reason for the warnings was that any access of the mysql.user view
will produce a warning if the mariadb.sys user does not exists.

Fixed with the following changes:
- Create the mariadb.sys user early, before any accesses to mysql.user.
- If run without --skip-grant, use GRANT to add new privileges in
  tables_priv. This ensures that the server knows about these privileges
  at once instead until the final FLUSH PRIVILEGES.